Move integration test to unit test
